fix signature and test for UTF-8 functions
On the advise of 8680, the unicode string was replaced by the hex
transcription of the corresponding codepoints (which he kindly
provided).
This might help to pass the overzealous encoding check of clang.

This commit also fixes a discrepancy between the signature of u8index in
declaration and definition.

+++ src/serialization/string_utils.cpp
@@ -1074,7 +1074,7 @@ utf8::string lowercase(const utf8::string& s)
 	return s;
 }
 
-unsigned int index(const utf8::string& str, const unsigned int index)
+size_t index(const utf8::string& str, const size_t index)
 {
 	// chr counts characters, i is the codepoint index
 	// remark: several functions rely on the fallback to str.length()
